foodie_be/requirements.txt
asgiref==3.8.1
Django==5.0.1
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
